fastapi/src/utils/langchain_utils.py
"""Minimal placeholder for Langchain utilities.

This module provides a stub implementation of the functions used by
the FastAPI service when generating embeddings or interacting with a
language model via LangChain.  Because the full LangChain library
and your fine‑tuned model are not included in this repository, the
functions here return dummy values to allow the API to run without
import errors.  You should replace these stubs with calls to your
actual embedding model or language chain when you integrate
advanced functionality.
"""

from typing import Optional, List


def get_open_ai_embeddings(text: str) -> Optional[List[float]]:
    """Return a dummy embedding for the provided text.

    In a full implementation this function would call OpenAI's
    embedding API or another provider to obtain a high‑dimensional
    vector representation of the text.  Here we simply return None
    to indicate that embeddings are unavailable.

    Args:
        text: The input text to embed.

    Returns:
        A list of floats representing the embedding, or None if
        embeddings are not available.
    """
    # Stub: return None or a small dummy vector.  You can replace this
    # with actual embedding logic as needed.
    return None